feat: create filtr chats finish and create helpers
